Add range requests for content downloads
Add a Seek() method to an internal adapter. This lets us use the
standard library http.ServeContent() to support range requests when
downloading content. At the moment, this change only applies to files
downloaded from our cache, and not those downloaded directly from tape.
But since we cache ALL files in production, and nothing is ever served
directly from tape, so this is moot.

This is part of the project to have end-to-end range requests for curate
downloads