Decode slots paths using the database
This patch moves the decoding into the server package where the database
is used to figure out the blob identifier directly rather than realizing
the entire item in memory and then walking the blob and version lists.
Items that have not been indexed into the database will be indexed upon
being accessed.

It was slightly odd that these server paths were defined in the item
package. At some point it would be good to completely remove the "Item
Cache" from the items package. It was a bit of a hack to get everything
working in the first place.