Add a token multiplexer
This is a first step to also supporting tokens being stored in the
database. The next step is to have the database code support the
TokenValidator interface, and then to have the startup code mux the
database and token files together. Eventually the token files could be
removed since they are hard to manage.